Extract and fix send tests
